pyproject.toml
[project]
name = "visa-attorney-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.82",
    "fastapi>=0.135.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.4",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
